sync-engine-shared - Retry connection / Offline events handling

+++ shared/sync-engine-shared/src/com/liferay/sync/engine/documentlibrary/event/BaseSyncDLObjectUpdateEvent.java
@@ -79,6 +79,10 @@ public class BaseSyncDLObjectUpdateEvent extends BaseEvent {
 			targetSyncFile.getRepositoryId(), getSyncAccountId(),
 			targetSyncFile.getTypePK());
 
+		if (sourceSyncFile == null) {
+			return;
+		}
+
 		sourceSyncFile.setUiEvent(SyncFile.UI_EVENT_TRASHED_REMOTE);
 
 		Files.deleteIfExists(Paths.get(sourceSyncFile.getFilePathName()));
