+++ shared/sync-engine-shared/src/com/liferay/sync/engine/filesystem/SyncWatchEventProcessor.java
@@ -14,8 +14,10 @@
 
 package com.liferay.sync.engine.filesystem;
 
+import com.liferay.sync.engine.model.SyncAccount;
 import com.liferay.sync.engine.model.SyncFile;
 import com.liferay.sync.engine.model.SyncWatchEvent;
+import com.liferay.sync.engine.service.SyncAccountService;
 import com.liferay.sync.engine.service.SyncFileService;
 import com.liferay.sync.engine.service.SyncWatchEventService;
 import com.liferay.sync.engine.util.FilePathNameUtil;
@@ -46,6 +48,13 @@ public class SyncWatchEventProcessor implements Runnable {
 			"eventType", true);
 
 		for (SyncWatchEvent syncWatchEvent : syncWatchEvents) {
+			SyncAccount syncAccount = SyncAccountService.fetchSyncAccount(
+				syncWatchEvent.getSyncAccountId());
+
+			if (syncAccount.getState() == SyncAccount.STATE_DISCONNECTED) {
+				continue;
+			}
+
 			if (_processedSyncWatchEventIds.contains(
 					syncWatchEvent.getSyncWatchEventId())) {
 
@@ -95,8 +104,13 @@ public class SyncWatchEventProcessor implements Runnable {
 				_logger.error(e.getMessage(), e);
 			}
 
-			SyncWatchEventService.deleteSyncWatchEvent(
-				syncWatchEvent.getSyncWatchEventId());
+			syncAccount = SyncAccountService.fetchSyncAccount(
+				syncWatchEvent.getSyncAccountId());
+
+			if (syncAccount.getState() != SyncAccount.STATE_DISCONNECTED) {
+				SyncWatchEventService.deleteSyncWatchEvent(
+					syncWatchEvent.getSyncWatchEventId());
+			}
 		}
 
 		_processedSyncWatchEventIds.clear();
@@ -146,8 +160,13 @@ public class SyncWatchEventProcessor implements Runnable {
 				syncWatchEvent.getSyncAccountId(), syncFile);
 		}
 
-		_processedSyncWatchEventIds.add(
-			relatedSyncWatchEvent.getSyncWatchEventId());
+		SyncAccount syncAccount = SyncAccountService.fetchSyncAccount(
+			syncWatchEvent.getSyncAccountId());
+
+		if (syncAccount.getState() != SyncAccount.STATE_DISCONNECTED) {
+			_processedSyncWatchEventIds.add(
+				relatedSyncWatchEvent.getSyncWatchEventId());
+		}
 	}
 
 	protected void addFolder(SyncWatchEvent syncWatchEvent) throws Exception {
